Removing more dead code.

+++ src/handler.c
@@ -173,7 +173,7 @@ static inline int handler_recv_parse(Handler *handler, HandlerParser *parser)
     check(rc == 1, "Failed to parse message from handler.");
 
     check(parser->target_count > 0, "Message sent had 0 targets: %.*s",
-            zmq_msg_size(inmsg), (char *)zmq_msg_data(inmsg));
+            (int)zmq_msg_size(inmsg), (char *)zmq_msg_data(inmsg));
 
     debug("Parsed message with %d targets, first: %d, uuid: %s, and body: %d",
             (int)parser->target_count, parser->targets[0],
